完成歷史紀錄之刪除、檢視之功能，以及修正一些分支功能(2020/3/20)

+++ app/src/main/java/com/jetec/usbmonitor/Controller/RecordHistoryActivity.kt
@@ -2,24 +2,27 @@ package com.jetec.usbmonitor.Controller
 
 import android.annotation.SuppressLint
 import android.app.Activity
+import android.app.AlertDialog
 import android.app.ProgressDialog
 import android.content.Context
+import android.content.Intent
 import android.graphics.Typeface
 import android.os.Bundle
+import android.os.SystemClock
 import android.util.Log
-import android.view.LayoutInflater
-import android.view.MenuItem
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import android.widget.Button
+import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
+import com.bumptech.glide.Glide
 import com.chauthai.swipereveallayout.SwipeRevealLayout
 import com.chauthai.swipereveallayout.ViewBinderHelper
+import com.github.clans.fab.FloatingActionMenu
 import com.jetec.usbmonitor.Model.CrashHandler
 import com.jetec.usbmonitor.Model.RoomDBHelper.Data
 import com.jetec.usbmonitor.Model.RoomDBHelper.DataBase
@@ -31,6 +34,10 @@ import org.json.JSONArray
 
 class RecordHistoryActivity : AppCompatActivity() {
     val TAG = RecordHistoryActivity::class.java.simpleName + "My"
+    companion object{
+        val RESULT_CODE = 1
+
+    }
 
     private lateinit var mAdapter: MyAdapter
 
@@ -39,28 +46,29 @@ class RecordHistoryActivity : AppCompatActivity() {
         setContentView(R.layout.activity_record_history)
         CrashHandler.getInstance().init(this)
         setMenu()
-        var dialog = ProgressDialog.show(this,"","讀取中...",true)
+
+        var dialog = ProgressDialog.show(this, "", "讀取中...", true)
         Thread {
             val savedData: MutableList<Data> = DataBase.getInstance(this).dataUao.allData
             val layoutManager = LinearLayoutManager(this)
             layoutManager.orientation = LinearLayoutManager.VERTICAL
             val recycler = findViewById<RecyclerView>(R.id.recyclerView_RecordHistoryDisplay)
             layoutManager.recycleChildrenOnDetach = true
-            recycler.addItemDecoration(DividerItemDecoration(this,DividerItemDecoration.VERTICAL))
+            recycler.addItemDecoration(DividerItemDecoration(this, DividerItemDecoration.VERTICAL))
             recycler.layoutManager = layoutManager
             dialog.setCancelable(true)
-            mAdapter = MyAdapter(savedData,this)
-
-
+            mAdapter = MyAdapter(savedData, this)
             runOnUiThread {
 
                 recycler.adapter = mAdapter
-                dialog.dismiss()
             }
+            SystemClock.sleep(1000)
+            dialog.dismiss()
 
         }.start()
 
     }
+
 
     /**設置標題列*/
     private fun setMenu() {
@@ -82,18 +90,27 @@ class RecordHistoryActivity : AppCompatActivity() {
         return super.onOptionsItemSelected(item)
     }
 
+    override fun onTouchEvent(event: MotionEvent?): Boolean {
+        val floatMenu = findViewById<FloatingActionMenu>(R.id.floatingActionMenuButton_Filter)
+        if (event!!.action >=0 && floatMenu.isOpened){
+            floatMenu.close(true)
+        }
+        return super.onTouchEvent(event)
+    }
+
     /**第一個RecyclerView*/
     private class MyAdapter : RecyclerView.Adapter<MyAdapter.ViewHolder> {
 
         val TAG = RecordHistoryActivity::class.java.simpleName + "My"
         private var data: MutableList<Data>
         private val binderHelper = ViewBinderHelper()
-        private val activity:Activity
+        private val activity: Activity
 
-        constructor(data: MutableList<Data>,activity: Activity) : super() {
+        constructor(data: MutableList<Data>, activity: Activity) : super() {
             this.data = data
             this.activity = activity
         }
+
 
         class ViewHolder(v: View) : RecyclerView.ViewHolder(v) {
             val tvTitle = v.findViewById<TextView>(R.id.textView_RH_Title)
@@ -105,7 +122,7 @@ class RecordHistoryActivity : AppCompatActivity() {
             val btReview = v.findViewById<Button>(R.id.button_RH_Review)
             val btModify = v.findViewById<Button>(R.id.button_RH_Modify)
             val swipeLayout = v.findViewById<SwipeRevealLayout>(R.id.swipe_layout)
-
+            val igPicture = v.findViewById<ImageView>(R.id.imageView_RH_Image)
             val parent = v
         }
 
@@ -123,7 +140,8 @@ class RecordHistoryActivity : AppCompatActivity() {
         override fun onBindViewHolder(holder: ViewHolder, position: Int) {
             setValue(holder, position)//打入需顯示的值
             setSecondRecyclerView(position, holder)//打入巢狀RecyclerView的值
-            setClick(holder, position,holder.parent.context)//設置點擊事件(們)
+            setClick(holder, position, holder.parent.context)//設置點擊事件(們)
+
         }
 
         private fun setClick(
@@ -132,22 +150,38 @@ class RecordHistoryActivity : AppCompatActivity() {
             context: Context
         ) {
             holder.btDelete.setOnClickListener {
-                val id = data[position].id
-                Thread{
-                    DataBase.getInstance(context).dataUao.deleteByID(id)
-                    data.removeAt(position)
-                    activity.runOnUiThread{
-                        notifyItemRemoved(position)
+                holder.swipeLayout.close(true)
+                val builder = AlertDialog.Builder(context)
+                builder.setMessage("Sure to delete this data?")
+                    .setPositiveButton(context.getString(R.string.oK_Button)) { dialog, i ->
+                        val id = data[position].id
+                        Thread {
+                            DataBase.getInstance(context).dataUao.deleteByID(id)
+                            data.removeAt(position)
+                            activity.runOnUiThread {
+                                notifyItemRemoved(position)
+                            }
+                        }.start()
                     }
-                }.start()
+                    .setNegativeButton(context.getString(R.string.cancelButton), null)
+                    .show()
+            }//onC1
 
-            }
-            holder.btModify.setOnClickListener {
-
-            }
             holder.btReview.setOnClickListener {
+                val intent = Intent(activity, ReviewDataActivity::class.java)
+                intent.putExtra("searchID", data[position].id)
+                activity.startActivity(intent)
+                holder.swipeLayout.close(true)
 
-            }
+            }//onC2
+            holder.btModify.setOnClickListener {
+                val id = data[position].id
+                val intent = Intent(activity,ModifyHistoryDataActivity::class.java)
+                intent.putExtra("position",id)
+                Log.d(TAG, ":GO ");
+                activity.startActivityForResult(intent,RESULT_CODE)
+
+            }//onC3
         }
 
         private fun setValue(
@@ -160,24 +194,33 @@ class RecordHistoryActivity : AppCompatActivity() {
             val context = holder.parent.context
             if (data[position].name.isEmpty()) {
 
-                title = context.getString(R.string.tester) +": " + holder.parent.context.getString(R.string.unfilled)
+                title =
+                    context.getString(R.string.tester) + ": " + holder.parent.context.getString(R.string.unfilled)
             } else {
-                title = context.getString(R.string.tester) +": "+ data[position].name
+                title = context.getString(R.string.tester) + ": " + data[position].name
             }
             val deviceID = context.getString(R.string.deviceId) + data[position].deviceUUID
             val deviceName = context.getString(R.string.deviceName) + data[position].deviceName
-            val time = context.getString(R.string.measureTime) + data[position].dateTime.replace("#", " ")
+            val time =
+                context.getString(R.string.measureTime) + data[position].dateTime.replace("#", " ")
 
             holder.tvTitle.text = title
             holder.tvID.text = deviceID
             holder.tvDeviceName.text = deviceName
             holder.tvTime.text = time
+
+            Glide.with(holder.parent).load(data[position].takeImage)
+                .centerCrop()
+                .placeholder(R.drawable.create_image)
+                .into(holder.igPicture)
+
         }
 
         private fun setSecondRecyclerView(
             position: Int,
             holder: ViewHolder
         ) {
+
             val jsonArray = JSONArray(data[position].json_MySave)
             val arrayList = ArrayList<HashMap<String, String>>()
             for (i in 0 until jsonArray.length()) {
@@ -196,19 +239,33 @@ class RecordHistoryActivity : AppCompatActivity() {
                 hashMap["Title"] = strTitle
                 hashMap["Origin"] = strOrigin
                 arrayList.add(hashMap)
+
+
+                val layoutManager =
+                    StaggeredGridLayoutManager(3, StaggeredGridLayoutManager.HORIZONTAL);
+                layoutManager.orientation = LinearLayoutManager.VERTICAL
+                val cAdapter = ChildRecyclerView(arrayList)
+                holder.recyclerChild.layoutManager = layoutManager
+
+                holder.recyclerChild.adapter = cAdapter
             }
 
-            val layoutManager = StaggeredGridLayoutManager(3, StaggeredGridLayoutManager.HORIZONTAL);
-            layoutManager.orientation = LinearLayoutManager.VERTICAL
-            val cAdapter = ChildRecyclerView(arrayList)
-            holder.recyclerChild.layoutManager = layoutManager
-            holder.recyclerChild.adapter = cAdapter
+
         }
 
 
     }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        Log.d(TAG, ":$requestCode $resultCode ${data?.getStringExtra("RESULT")} ");
+        if (requestCode == RESULT_CODE&&resultCode == 1){
+
+        }
+    }
+
     /**包在第一個RecyclerView內的那個Recycler*/
-    private class ChildRecyclerView(val recordInfo:ArrayList<HashMap<String,String>>) :
+    private class ChildRecyclerView(val recordInfo: ArrayList<HashMap<String, String>>) :
         RecyclerView.Adapter<ChildRecyclerView.ViewHolder>() {
 
         val TAG = RecordHistoryActivity::class.java.simpleName + "My"
@@ -234,11 +291,17 @@ class RecordHistoryActivity : AppCompatActivity() {
 
         @SuppressLint("SetTextI18n")
         override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            var unit = Tools.setUnit(Tools.hex2Dec(recordInfo[position]["Origin"]!!.substring(12,14)).toInt())
-            holder.tvTitle.text ="${recordInfo[position]["Title"]}:\n${recordInfo[position]["value"]+unit}"
-            holder.tvPV.text = holder.parent.context.getString(R.string.PV)+": "+recordInfo[position]["PV"]
-            holder.tvEH.text =holder.parent.context.getString(R.string.EH)+": "+ recordInfo[position]["EH"]
-            holder.tvEL.text =holder.parent.context.getString(R.string.EL) +": "+recordInfo[position]["EL"]
+            var unit = Tools.setUnit(
+                Tools.hex2Dec(recordInfo[position]["Origin"]!!.substring(12, 14)).toInt()
+            )
+            holder.tvTitle.text =
+                "${recordInfo[position]["Title"]}:\n${recordInfo[position]["value"] + unit}"
+            holder.tvPV.text =
+                holder.parent.context.getString(R.string.PV) + ": " + recordInfo[position]["PV"]
+            holder.tvEH.text =
+                holder.parent.context.getString(R.string.EH) + ": " + recordInfo[position]["EH"]
+            holder.tvEL.text =
+                holder.parent.context.getString(R.string.EL) + ": " + recordInfo[position]["EL"]
 
 
         }
